# neus-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi Specification ",
        "description": "OpenApi Documentation for Alem Equb",
        "termsOfService": "term of service",
        "contact": {
            "name": "Habtamu",
            "email": "habtamuadugna.d@gmail.com"
        },
        "license": {
            "name": "Licence name ",
            "url": "http://url licence"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "Local ENV"
        }
    ],
    "security": [
        {
            "keycloak": []
        }
    ],
    "paths": {
...
        "/exams": {
            "get": {
                "tags": [
                    "exams"
                ],
                "operationId": "getPageOfExams",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseExamDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "exams"
                ],
                "operationId": "updateExam",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateExamDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "exams"
                ],
                "operationId": "createExam",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateExamDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/exams/{exam-id}": {
            "get": {
                "tags": [
                    "exams"
                ],
                "operationId": "getExamById",
                "parameters": [
                    {
                        "name": "exam-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "exams"
                ],
                "operationId": "deleteExam",
                "parameters": [
                    {
                        "name": "exam-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/exams/years/{exam-type}": {
            "get": {
                "tags": [
                    "exams"
                ],
                "operationId": "getYearsByExamType",
                "parameters": [
                    {
                        "name": "exam-type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "TEST",
                                "ERMP",
                                "USMLE_STEP_1",
                                "USMLE_STEP_2"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NumberDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/exams/list/{exam-type}": {
            "get": {
                "tags": [
                    "exams"
                ],
                "operationId": "getExamNamesByExamType",
                "parameters": [
                    {
                        "name": "exam-type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "TEST",
                                "ERMP",
                                "USMLE_STEP_1",
                                "USMLE_STEP_2"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamNameDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/exams/list/{exam-type}/{year}": {
            "get": {
                "tags": [
                    "exams"
                ],
                "operationId": "getExamNamesByExamTypeAndYear",
                "parameters": [
                    {
                        "name": "exam-type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "TEST",
                                "ERMP",
                                "USMLE_STEP_1",
                                "USMLE_STEP_2"
                            ]
                        }
                    },
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExamNameDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/exams/detail/{exam-id}": {
            "get": {
                "tags": [
                    "exams"
                ],
                "operationId": "getExamDetail",
                "parameters": [
                    {
                        "name": "exam-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "block",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExamDetailDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateExamDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "examType": {
                        "type": "string",
                        "enum": [
                            "TEST",
                            "ERMP",
                            "USMLE_STEP_1",
                            "USMLE_STEP_2"
                        ]
                    },
                    "year": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "randomQuestionCount": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "required": [
                    "duration",
                    "examType",
                    "requiredSubLevel",
                    "year"
                ]
            },
...
            "ExamDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "examType": {
                        "type": "string",
                        "enum": [
                            "TEST",
                            "ERMP",
                            "USMLE_STEP_1",
                            "USMLE_STEP_2"
                        ]
                    },
                    "year": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "numberOfQuestions": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "randomQuestionCount": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "PageResponseExamDto": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExamDto"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "NumberDto": {
                "type": "object",
                "properties": {
                    "value": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ExamNameDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ChoiceDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    },
                    "isCorrect": {
                        "type": "boolean"
                    },
                    "percentage": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
            "ExamDetailDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "requiredSubLevel": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "BASIC",
                            "ADVANCED",
                            "PREMIUM"
                        ]
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuestionDetailDto"
                        }
                    }
                }
            },
            "QuestionDetailDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "examId": {
                        "type": "string"
                    },
                    "questionNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "questionText": {
                        "type": "string"
                    },
                    "choices": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ChoiceDto"
                        }
                    },
                    "totalResponses": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "explanation": {
                        "type": "string"
                    },
                    "imgUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "department": {
                        "type": "string",
                        "enum": [
                            "Anatomy",
                            "Internal_Medicine",
                            "Pediatrics",
                            "OBGYN",
                            "Surgery",
                            "Medical_Ethics",
                            "Dermatology",
                            "Ophthalmology",
                            "Psychiatry",
                            "ENT",
                            "Pathology",
                            "Medical_Radiology",
                            "Anesthesiology",
                            "Behavioral_Science",
                            "Biochemistry",
                            "Biostatics",
                            "Embryology",
                            "Genetics",
                            "Histology",
                            "Immunology",
                            "Microbiology",
                            "Pathophysiology",
                            "Pharmacology",
                            "Physiology"
                        ]
                    },
                    "blockNumber": {
                        "type": "string",
                        "enum": [
                            "BLOCK_01",
                            "BLOCK_02",
                            "BLOCK_03",
                            "BLOCK_04",
                            "BLOCK_05",
                            "BLOCK_06",
                            "BLOCK_07",
                            "BLOCK_08",
                            "BLOCK_09",
                            "BLOCK_10"
                        ]
                    }
                }
            }
        },
        "securitySchemes": {
            "keycloak": {
                "type": "oauth2",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:9090/realms/neus/protocol/openid-connect/auth",
                        "tokenUrl": "http://localhost:9090/realms/neus/protocol/openid-connect/token",
                        "scopes": {}
                    }
                }
            }
        }
    }
}